8293121: Absorb UnixCopyFile into UnixFileSystem; remove UnixCopyFile hierarchy
